Separate "span" into its own package (#1141)
