+++ model/span/_meta/fields.yml
@@ -1,0 +1,59 @@
+- key: apm-span
+  title: APM Span
+  description: Span-specific data for APM.
+  fields:
+    - name: view spans
+      format: url
+      label_template: "View Spans"
+
+    - name: span
+      type: group
+      dynamic: false
+      fields:
+
+        - name: id
+          type: long
+          description: >
+            A locally unique ID of the span.
+
+        - name: name
+          type: keyword
+          count: 1
+          description: >
+            Generic designation of a span in the scope of a transaction.
+
+        - name: type
+          type: keyword
+          count: 1
+          description: >
+            Keyword of specific relevance in the service's domain (eg: 'db.postgresql.query', 'template.erb', 'cache', etc).
+
+        - name: start
+          type: group
+          description:
+          fields:
+            - name: us
+              type: long
+              count: 1
+              description: >
+                Offset relative to the transaction's timestamp identifying the start of the span, in microseconds.
+
+        - name: duration
+          type: group
+          description:
+          fields:
+
+            - name: us
+              type: long
+              count: 1
+              description: >
+                Duration of the span, in microseconds.
+              format: duration
+              input_format: microseconds
+              output_format: asMilliseconds
+              output_precision: 0
+
+        - name: parent
+          type: long
+          description: >
+             The locally unique ID of the parent of the span.
