+++ model/span/span_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package transaction
+package span
 
 import (
 	"errors"
